+++ mov2mov/__init__.py
@@ -1,0 +1,3 @@
+# from processing import StableDiffusionProcessingMov2Mov
+#
+# __all__ = ["StableDiffusionProcessingMov2Mov"]
